Move ColorListItem to `@/components/layout/BaseColors`
