+++ src/components/layout/BaseColors/index.ts
@@ -1,0 +1,1 @@
+export * from './ColorListItem';
